[Bronze IV] Title: 영수증, Time: 68 ms, Memory: 5816 KB -BaekjoonHub

+++ 백준/Bronze/25304. 영수증/영수증.cs
@@ -4,7 +4,7 @@ long sum = 0;
 string str = "No";
 for(int i = 0; i < num; i++){
     string[] str1 = Console.ReadLine().Split();
-    sum += int.Parse(str1[0]) * int.Parse(str1[1]);
+    sum += long.Parse(str1[0]) * long.Parse(str1[1]);
 }
 if(price == sum)
     str = "Yes";
